+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 
 ### Added
+- LinkHandles for `path` and `polygon` elements
 
 ## [0.0.1] - 2019-02-15
 ### Added